Updated styling, Product model and view; added form helpers to product index view

+++ db/migrate/20170425003631_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[5.0]
   def change
     create_table :products do |t|
       t.string :product_name
-      t.decimal :product_price
+      t.integer :product_price
       t.string :product_image
       t.text :product_description
